+++ sample-app/src/main/scala/example/model/concert/actor/ConcertActorBehaviorFactory.scala
@@ -3,7 +3,7 @@ package example.model.concert.actor
 import akka.actor.typed.Behavior
 import akka.persistence.typed.PersistenceId
 import example.model.concert.ConcertId
-import example.model.concert.actor.ConcertActorProtocol.ConcertCommandRequest
+import example.model.concert.actor.ConcertActor.ConcertCommandRequest
 
 trait ConcertActorBehaviorFactory {
   def apply(id: ConcertId, persistenceId: PersistenceId): Behavior[ConcertCommandRequest]
