+++ sample-app/src/main/scala/example/model/concert/actor/DefaultConcertActorWithEventPersistence.scala
@@ -7,7 +7,7 @@ import example.model.KryoSerializable
 import example.model.concert.ConcertError._
 import example.model.concert.ConcertEvent._
 import example.model.concert._
-import example.model.concert.actor.ConcertActorProtocol._
+import example.model.concert.actor.ConcertActor._
 
 import java.time.ZonedDateTime
