+++ sample-app/src/main/scala/example/model/concert/service/BoxOfficeService.scala
@@ -1,7 +1,7 @@
 package example.model.concert.service
 
 import example.model.concert.ConcertId
-import example.model.concert.actor.ConcertActorProtocol._
+import example.model.concert.actor.ConcertActor._
 
 import scala.concurrent.Future
