+++ sample-app/src/main/scala/example/model/concert/service/MyBoxOfficeService.scala
@@ -14,7 +14,7 @@ final class MyBoxOfficeService(
     system: ActorSystem[Nothing],
     behaviorFactory: ConcertActorBehaviorFactory,
 ) extends BoxOfficeService {
-  import example.model.concert.actor.ConcertActorProtocol._
+  import example.model.concert.actor.ConcertActor._
 
   // 設定を読み込む
   private val config                            = BoxOfficeServiceConfig(system)
